+++ components/search_provider_logos/logo_cache.cc
@@ -7,14 +7,16 @@
 #include <stddef.h>
 #include <stdint.h>
 
+#include <memory>
 #include <utility>
 
 #include "base/files/file_util.h"
 #include "base/json/json_reader.h"
 #include "base/json/json_writer.h"
-#include "base/memory/ptr_util.h"
 #include "base/strings/string_number_conversions.h"
 #include "base/values.h"
+
+namespace search_provider_logos {
 
 namespace {
 
@@ -23,11 +25,28 @@ const char kSourceUrlKey[] = "url";
 const char kExpirationTimeKey[] = "expiration_time";
 const char kCanShowAfterExpirationKey[] = "can_show_after_expiration";
 const char kFingerprintKey[] = "fingerprint";
+const char kTypeKey[] = "type";
 const char kOnClickURLKey[] = "on_click_url";
+const char kFullPageURLKey[] = "full_page_url";
 const char kAltTextKey[] = "alt_text";
 const char kMimeTypeKey[] = "mime_type";
 const char kNumBytesKey[] = "num_bytes";
 const char kAnimatedUrlKey[] = "animated_url";
+const char kLogUrlKey[] = "log_url";
+const char kCtaLogUrlKey[] = "cta_log_url";
+const char kShortLinkKey[] = "short_link";
+const char kIframeWidthPx[] = "iframe_width_px";
+const char kIframeHeightPx[] = "iframe_height_px";
+
+const char kShareButtonX[] = "share_button_x";
+const char kShareButtonY[] = "share_button_y";
+const char kShareButtonOpacity[] = "share_button_opacity";
+const char kShareButtonIcon[] = "share_button_icon";
+const char kShareButtonBg[] = "share_button_bg";
+
+const char kSimpleType[] = "SIMPLE";
+const char kAnimatedType[] = "ANIMATED";
+const char kInteractiveType[] = "INTERACTIVE";
 
 bool GetTimeValue(const base::DictionaryValue& dict,
                   const std::string& key,
@@ -46,70 +65,101 @@ void SetTimeValue(base::DictionaryValue& dict,
                   const std::string& key,
                   const base::Time& time) {
   int64_t internal_time_value = time.ToInternalValue();
-  dict.SetString(key, base::Int64ToString(internal_time_value));
+  dict.SetString(key, base::NumberToString(internal_time_value));
+}
+
+LogoType LogoTypeFromString(base::StringPiece type) {
+  if (type == kSimpleType) {
+    return LogoType::SIMPLE;
+  }
+  if (type == kAnimatedType) {
+    return LogoType::ANIMATED;
+  }
+  if (type == kInteractiveType) {
+    return LogoType::INTERACTIVE;
+  }
+  LOG(WARNING) << "invalid type " << type;
+  return LogoType::SIMPLE;
+}
+
+std::string LogoTypeToString(LogoType type) {
+  switch (type) {
+    case LogoType::SIMPLE:
+      return kSimpleType;
+    case LogoType::ANIMATED:
+      return kAnimatedType;
+    case LogoType::INTERACTIVE:
+      return kInteractiveType;
+  }
+  NOTREACHED();
+  return "";
 }
 
 }  // namespace
-
-namespace search_provider_logos {
 
 LogoCache::LogoCache(const base::FilePath& cache_directory)
     : cache_directory_(cache_directory),
       metadata_is_valid_(false) {
   // The LogoCache can be constructed on any thread, as long as it's used
-  // on a single thread after construction.
-  thread_checker_.DetachFromThread();
+  // on a single sequence after construction.
+  DETACH_FROM_SEQUENCE(sequence_checker_);
 }
 
 LogoCache::~LogoCache() {
-  DCHECK(thread_checker_.CalledOnValidThread());
+  DCHECK_CALLED_ON_VALID_SEQUENCE(sequence_checker_);
 }
 
 void LogoCache::UpdateCachedLogoMetadata(const LogoMetadata& metadata) {
-  DCHECK(thread_checker_.CalledOnValidThread());
+  DCHECK_CALLED_ON_VALID_SEQUENCE(sequence_checker_);
   DCHECK(metadata_);
   DCHECK_EQ(metadata_->fingerprint, metadata.fingerprint);
 
-  UpdateMetadata(base::WrapUnique(new LogoMetadata(metadata)));
+  UpdateMetadata(std::make_unique<LogoMetadata>(metadata));
   WriteMetadata();
 }
 
 const LogoMetadata* LogoCache::GetCachedLogoMetadata() {
-  DCHECK(thread_checker_.CalledOnValidThread());
+  DCHECK_CALLED_ON_VALID_SEQUENCE(sequence_checker_);
   ReadMetadataIfNeeded();
   return metadata_.get();
 }
 
 void LogoCache::SetCachedLogo(const EncodedLogo* logo) {
-  DCHECK(thread_checker_.CalledOnValidThread());
-  std::unique_ptr<LogoMetadata> metadata;
-  if (logo) {
-    metadata.reset(new LogoMetadata(logo->metadata));
-    logo_num_bytes_ = static_cast<int>(logo->encoded_image->size());
+  DCHECK_CALLED_ON_VALID_SEQUENCE(sequence_checker_);
+  if (!logo) {
+    UpdateMetadata(nullptr);
+    DeleteLogoAndMetadata();
+    return;
   }
-  UpdateMetadata(std::move(metadata));
-  WriteLogo(logo ? logo->encoded_image : NULL);
+
+  logo_num_bytes_ =
+      logo->encoded_image ? static_cast<int>(logo->encoded_image->size()) : 0;
+  UpdateMetadata(std::make_unique<LogoMetadata>(logo->metadata));
+  WriteLogo(logo->encoded_image);
 }
 
 std::unique_ptr<EncodedLogo> LogoCache::GetCachedLogo() {
-  DCHECK(thread_checker_.CalledOnValidThread());
+  DCHECK_CALLED_ON_VALID_SEQUENCE(sequence_checker_);
 
   ReadMetadataIfNeeded();
   if (!metadata_)
     return nullptr;
 
-  scoped_refptr<base::RefCountedString> encoded_image =
-      new base::RefCountedString();
-  if (!base::ReadFileToString(GetLogoPath(), &encoded_image->data())) {
-    UpdateMetadata(nullptr);
-    return nullptr;
-  }
+  scoped_refptr<base::RefCountedString> encoded_image;
+  if (logo_num_bytes_ != 0) {
+    encoded_image = new base::RefCountedString();
 
-  if (encoded_image->size() != static_cast<size_t>(logo_num_bytes_)) {
-    // Delete corrupt metadata and logo.
-    DeleteLogoAndMetadata();
-    UpdateMetadata(nullptr);
-    return nullptr;
+    if (!base::ReadFileToString(GetLogoPath(), &encoded_image->data())) {
+      UpdateMetadata(nullptr);
+      return nullptr;
+    }
+
+    if (encoded_image->size() != static_cast<size_t>(logo_num_bytes_)) {
+      // Delete corrupt metadata and logo.
+      DeleteLogoAndMetadata();
+      UpdateMetadata(nullptr);
+      return nullptr;
+    }
   }
 
   std::unique_ptr<EncodedLogo> logo(new EncodedLogo());
@@ -122,24 +172,52 @@ std::unique_ptr<EncodedLogo> LogoCache::GetCachedLogo() {
 std::unique_ptr<LogoMetadata> LogoCache::LogoMetadataFromString(
     const std::string& str,
     int* logo_num_bytes) {
-  std::unique_ptr<base::Value> value = base::JSONReader::Read(str);
+  std::unique_ptr<base::Value> value = base::JSONReader::ReadDeprecated(str);
   base::DictionaryValue* dict;
   if (!value || !value->GetAsDictionary(&dict))
     return nullptr;
 
   std::unique_ptr<LogoMetadata> metadata(new LogoMetadata());
-  if (!dict->GetString(kSourceUrlKey, &metadata->source_url) ||
+  std::string source_url;
+  std::string type;
+  std::string on_click_url;
+  std::string full_page_url;
+  std::string animated_url;
+  std::string log_url;
+  std::string cta_log_url;
+  std::string short_link;
+  if (!dict->GetString(kSourceUrlKey, &source_url) ||
       !dict->GetString(kFingerprintKey, &metadata->fingerprint) ||
-      !dict->GetString(kOnClickURLKey, &metadata->on_click_url) ||
+      !dict->GetString(kTypeKey, &type) ||
+      !dict->GetString(kOnClickURLKey, &on_click_url) ||
+      !dict->GetString(kFullPageURLKey, &full_page_url) ||
       !dict->GetString(kAltTextKey, &metadata->alt_text) ||
-      !dict->GetString(kAnimatedUrlKey, &metadata->animated_url) ||
+      !dict->GetString(kAnimatedUrlKey, &animated_url) ||
+      !dict->GetString(kLogUrlKey, &log_url) ||
+      !dict->GetString(kCtaLogUrlKey, &cta_log_url) ||
+      !dict->GetString(kShortLinkKey, &short_link) ||
       !dict->GetString(kMimeTypeKey, &metadata->mime_type) ||
       !dict->GetBoolean(kCanShowAfterExpirationKey,
                         &metadata->can_show_after_expiration) ||
       !dict->GetInteger(kNumBytesKey, logo_num_bytes) ||
+      !dict->GetInteger(kShareButtonX, &metadata->share_button_x) ||
+      !dict->GetInteger(kShareButtonY, &metadata->share_button_y) ||
+      !dict->GetDouble(kShareButtonOpacity, &metadata->share_button_opacity) ||
+      !dict->GetString(kShareButtonIcon, &metadata->share_button_icon) ||
+      !dict->GetString(kShareButtonBg, &metadata->share_button_bg) ||
+      !dict->GetInteger(kIframeWidthPx, &metadata->iframe_width_px) ||
+      !dict->GetInteger(kIframeHeightPx, &metadata->iframe_height_px) ||
       !GetTimeValue(*dict, kExpirationTimeKey, &metadata->expiration_time)) {
     return nullptr;
   }
+  metadata->type = LogoTypeFromString(type);
+  metadata->source_url = GURL(source_url);
+  metadata->on_click_url = GURL(on_click_url);
+  metadata->full_page_url = GURL(full_page_url);
+  metadata->animated_url = GURL(animated_url);
+  metadata->log_url = GURL(log_url);
+  metadata->cta_log_url = GURL(cta_log_url);
+  metadata->short_link = GURL(short_link);
 
   return metadata;
 }
@@ -149,15 +227,27 @@ void LogoCache::LogoMetadataToString(const LogoMetadata& metadata,
                                      int num_bytes,
                                      std::string* str) {
   base::DictionaryValue dict;
-  dict.SetString(kSourceUrlKey, metadata.source_url);
+  dict.SetString(kSourceUrlKey, metadata.source_url.spec());
   dict.SetString(kFingerprintKey, metadata.fingerprint);
-  dict.SetString(kOnClickURLKey, metadata.on_click_url);
+  dict.SetString(kTypeKey, LogoTypeToString(metadata.type));
+  dict.SetString(kOnClickURLKey, metadata.on_click_url.spec());
+  dict.SetString(kFullPageURLKey, metadata.full_page_url.spec());
   dict.SetString(kAltTextKey, metadata.alt_text);
-  dict.SetString(kAnimatedUrlKey, metadata.animated_url);
+  dict.SetString(kAnimatedUrlKey, metadata.animated_url.spec());
+  dict.SetString(kLogUrlKey, metadata.log_url.spec());
+  dict.SetString(kCtaLogUrlKey, metadata.cta_log_url.spec());
+  dict.SetString(kShortLinkKey, metadata.short_link.spec());
   dict.SetString(kMimeTypeKey, metadata.mime_type);
   dict.SetBoolean(kCanShowAfterExpirationKey,
                   metadata.can_show_after_expiration);
   dict.SetInteger(kNumBytesKey, num_bytes);
+  dict.SetInteger(kShareButtonX, metadata.share_button_x);
+  dict.SetInteger(kShareButtonY, metadata.share_button_y);
+  dict.SetDouble(kShareButtonOpacity, metadata.share_button_opacity);
+  dict.SetString(kShareButtonIcon, metadata.share_button_icon);
+  dict.SetString(kShareButtonBg, metadata.share_button_bg);
+  dict.SetInteger(kIframeWidthPx, metadata.iframe_width_px);
+  dict.SetInteger(kIframeHeightPx, metadata.iframe_height_px);
   SetTimeValue(dict, kExpirationTimeKey, metadata.expiration_time);
   base::JSONWriter::Write(dict, str);
 }
@@ -206,7 +296,7 @@ void LogoCache::WriteLogo(scoped_refptr<base::RefCountedMemory> encoded_image) {
   if (!EnsureCacheDirectoryExists())
     return;
 
-  if (!metadata_ || !encoded_image.get()) {
+  if (!metadata_) {
     DeleteLogoAndMetadata();
     return;
   }
@@ -220,10 +310,9 @@ void LogoCache::WriteLogo(scoped_refptr<base::RefCountedMemory> encoded_image) {
   if (!base::DeleteFile(metadata_path, false))
     return;
 
-  if (base::WriteFile(
-          logo_path,
-          encoded_image->front_as<char>(),
-          static_cast<int>(encoded_image->size())) == -1) {
+  if (encoded_image &&
+      base::WriteFile(logo_path, encoded_image->front_as<char>(),
+                      static_cast<int>(encoded_image->size())) == -1) {
     base::DeleteFile(logo_path, false);
     return;
   }
